Cleanup

+++ Systems/RenderSystem.cpp
@@ -12,8 +12,6 @@
 void RenderSystem::run(World& world) {
 	if (world.components_exist_of_type<Sprite>()) {
 		for (auto* spr : world.get_components<Sprite>()) {
-		//for (size_t i = 0; i < world.get_components<Sprite>().size(); i++) {
-			//Sprite* spr = world.get_components<Sprite>()[i];
 			render_sprite(spr, spr->get_owner_component<Transform>(), world);
 		}
 	}
@@ -27,7 +25,7 @@ void RenderSystem::render_sprite(Sprite* spr, Transform* tr, World& world) {
 		Vector2 scale = tr->get_scale();
 		Vector2I size = spr->sprite_size();
 
-		SDL_Rect pos_rect{static_cast<int>(pos.x), static_cast<int>(pos.y), size.x * scale.x, size.y * scale.y};
+		SDL_Rect pos_rect{static_cast<int>(pos.x), static_cast<int>(pos.y), static_cast<int>(size.x * scale.x), static_cast<int>(size.y * scale.y)};
 
 		SDL_RenderCopyEx(Window::get_renderer(), spr->get_texture(), nullptr, &pos_rect, rot, nullptr, SDL_RendererFlip::SDL_FLIP_NONE);
 	}
